Added network propagation
