+++ sync/sync.js
@@ -2,6 +2,9 @@ const { getOwnSyncState, setReceivedSyncData } = require( "./syncutil.js" )
 const accounts = require( "../shared/accounts.js" )
 const { GameServer, GetGameServers, AddGameServer, UpdateGameServer } = require( "../shared/gameserver.js" )
 const { startSync } = require( "./broadcast.js" )
+const { addNetworkNode, getNetworkNodes } = require("./network.js")
+const { getInstanceById } = require("./util.js")
+
 module.exports = {
 	// eventName: async (data) => {
 	//     try { 
@@ -106,6 +109,15 @@ module.exports = {
 		catch( e )
 		{
 			if( process.env.USE_DATASYNC_LOGGING ) console.log( e )
+		}
+	},
+	addNetworkNode: async (data) => {
+		let payload = data.payload
+		try { 
+			addNetworkNode(payload.id, payload.host, (await getInstanceById(payload.id)).port, payload.token)
+			console.log("Updated network! Current network: " + Object.keys(await getNetworkNodes()))
+		} catch(e) {
+			if(process.env.USE_DATASYNC_LOGGING) console.log(e)
 		}
 	}
 }
