+++ test.env
@@ -2,7 +2,7 @@
 USE_HTTPS=
 USE_FASTIFY_LOGGER=
 LISTEN_IP=0.0.0.0
-LISTEN_PORT=8080
+LISTEN_PORT=8081
 REQUIRE_SESSION_TOKEN=
 ALLOW_NON_MODINFO=
 
@@ -20,4 +20,3 @@ USE_DATASYNC_LOGGING=1
 DATASYNC_OWN_ID=B
 ENCRYPT_ALGO=
 DATASYNC_TOKEN_BYTES=
-NETWORK_EXTERNAL_IP=localhost
